refactor construction types
